Remove search input; refactor for simplicity; add comments
