--- ui/src/row/StockRowTrait.ts
@@ -1,7 +1,0 @@
-import { RowTrait } from "@swim/table";
-
-export class StockRowTrait extends RowTrait {
-  constructor() {
-    super();
-  }
-}
