--- ui/src/row/StockRowView.ts
@@ -1,7 +1,0 @@
-import { RowView } from "@swim/table";
-
-export class StockRowView extends RowView {
-  constructor(node: HTMLElement) {
-    super(node);
-  }
-}
