--- ui/src/table/index.ts
@@ -1,1 +1,0 @@
-export * from "./StockTableController";
